+++ changelog.md
@@ -1,3 +1,12 @@
+### 0.9.6
+
+- bump bioimageio.spec library version to 0.5.7.4
+- increase default reprducibility tolerance
+- unify quantile (vs percentile) variable names
+- add quantile computation method parameter
+- accept `SampleQuantile` or `DatasetQuantile` as `min`/`max` arguments to `proc_ops.Clip`
+- save actual output during model testing only if an explicit working directory was specified to produce less clutter
+
 ### 0.9.5
 
 - bump bioimageio.spec library version to 0.5.6.0